contract changes and arbitration action fix

+++ backend/demux-server/src/services/demux/updaters/arb/deleteCase.js
@@ -6,8 +6,16 @@ async function deleteCaseHandler (state, payload, blockInfo, context) {
         let case_id = parseInt(payload.data.case_id);
 
         // Delete|Remove CaseFile
-        await state.case.findOneAndDelete({ case_id: case_id }).exec();
 
+        let caseState = await state.case.findOne({ case_id: case_id }).exec();
+        let accepted_claims;
+        if (caseState) {
+            ({ accepted_claims } = caseState);
+            for ( let claim_id of accepted_claims ) {
+                await state.claim.findOneAndDelete({ claim_id: claim_id }).exec();
+            }
+            await state.case.findOneAndDelete({ case_id: case_id }).exec();
+        }
     } catch (err) {
         console.error('DeleteCase updater error: ', err);
     }
